Introduce the ability to fix the Gopkg.lock file
It will update the Gopkg.lock file with the appropriate revisions, run
`dep ensure` to ensure compatibility, and then rerun the check to make
sure nothing changed.

This now uses a vendored copy of the same toml library that dep uses
because dep does not update a `Gopkg.lock` file that hasn't been changed
even if the bytes are different.